Task 2
Switched alot of the items which were in board.java and moved them to checkers.java and inctances of Board and Square are created there

In addition, for task 2 only the white squares will have pieces on it.  So through the checkers class the piece is worked out then is passed to the Square constructor. These are then loaded onto the board.