Task3
Moved a lot of the things which used to be in the Checkers class to the Board class. The checkers class is used to just generate a version of itself which is used within generating the board

Inside the Board constructor is also where the Squares are generated.

For the action Listener, it uses an if statement and a boolean oriented if statements to differentiate the first and second click.

The first click will store the tile which was clicked.
The second click is where the moveTo function is used and uses mutators to switch the piece on the tile you first clicked with the tile you second clicked.
In the moveTo function it makes sure that the move being made is llegal too