Task 4
I have updated the Square constructor so that when there is a white square and is in yLoc 0-2 then the red piece will be shown there.

To make the code modular I have only changed the direction in which the pieces are able to go in. This is found out through the change integer changing to +1(going down the board) and -1(going up the board.

I then made sure that the pieces were only allowed to move to an empty tile. This was done through a simple if statement.