Task 5
I added the canMoveTo function in square which checked if a piece can be moved to a certain tile. This is a function which makes sure if the rules of the game are being followed.

To highlight the squares I made two functions. One in the board class called showOption and another called isWhite in the class Square.
The isWhite function returns the direction in which pieces are meant to move.
The showOption function then will make the available move options yellow on the first selection of a piece.
On the second click(when the user makes a move or doesn't make a move) then gets rid of the yellow option panes in which the user can move.

Using the canMoveTo I have made it so that all rules (for non-king pieces) to do with movement are followed.