added canJumpTo method and removed bugs
added an if statement to the canMoveTo function to avoid a bug where the piece was able to jump over its own piece. I also added a further if statement to avoid a bug where the highlighting would delete a piece if there was another piece inbetween it.

I changed the actionListener so that a jump would be performed if the target piece is yellow. It would then remove the yellow pieces after rather than before. This was used to remove additional proccesses

I made an additional method called canJumpTo which would show if a piece could jump to a location. I did this so later would be able to add certain features such as forced jump.

I also removed redundant if statements and added additional comments so another user will be able to know whats going on in the code.