Forced Jump (a)
I've made it so that upon selecting a tile, it forces it to make a jump even if it can make a normal move to another location.

I have also added new images which will be used to show when the piece (which changes image) has a jump available so the jump move will be forced.

I need to add the proper forced jump so that a piece with a double jump available has to be used.