Forced Jump (b)
Added a new function called checkPossibleJump which will check if there is a jump which is possible on the board.

I need to still refine it working and where to put the checkPossibleJump in the code.

The method will scan the entire board and highlight which draughs have to make a jump and force the user to select them.